✨ feat: Adiciona controlador SocketIO     Implementa a classe Controller para gerenciar a integração do SocketIO com o Quart, incluindo configuração de servidor, gerenciamento de sessões e manipulação de eventos.
Arquivos Alterados:
    - quart_socketio/core.py: Criação do controlador
    - quart_socketio/main.py: Implementação do SocketIO
